feat: add not found page, add routes
