+++ class-components/src/App.css
@@ -4,7 +4,7 @@
   padding: 2rem;
   text-align: center;
 }
-
+/* 
 .app {
   display: flex;
   flex-direction: column;
@@ -34,4 +34,4 @@
 .loading {
   font-size: 2rem;
   font-weight: 700;
-}
+} */
