+++ class-components/src/components/Results/Results.tsx
@@ -2,7 +2,6 @@ import './Results.css';
 import { IResultsProps } from '../../types/SearchTypes';
 
 const Results = ({ results }: IResultsProps) => {
-  console.log(results);
   return results.length ? (
     <div className="result">
       {results.map(result => (
