+++ class-components/src/constants.ts
@@ -1,2 +1,7 @@
 export const LS_QUERY = 'searchQuery';
 export const BASE_URL = 'https://swapi.dev/api/';
+
+export const ROUTES = {
+  HOME_PAGE: '/',
+  NOT_FOUND_PAGE: '*',
+};
