+++ class-components/src/pages/Home/Home.css
@@ -1,0 +1,30 @@
+.app {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid #000;
+  border-radius: 10px;
+  padding: 10px 20px;
+}
+
+.search-field {
+  flex: 0.5;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+}
+
+.results-field {
+  flex: 3;
+  overflow-y: auto;
+  padding: 20px;
+}
+
+.error-btn {
+  font-size: 18px;
+}
+
+.loading {
+  font-size: 2rem;
+  font-weight: 700;
+}
