--- class-components/src/pages/Main/Main.tsx
@@ -1,15 +1,0 @@
-// import React from 'react'
-
-// const Main = () => {
-//   return (
-//       <div className="results-field">
-//         {isLoading ? (
-//           <div className="loading">Loading...</div>
-//         ) : (
-//           <Results results={results} />
-//         )}
-//       </div>
-//   )
-// }
-
-// export default Main;
