+++ class-components/src/pages/Not_Found/Not_Found.tsx
@@ -1,0 +1,15 @@
+import NotFoundButton from '../../components/UI/NotFoundButton/NotFoundButton';
+import './Not_Found.css';
+
+const Not_Found = () => {
+  return (
+    <div className="not-found__page">
+      <div className="not-found__content">
+        <h1>Page not found</h1>
+        <NotFoundButton />
+      </div>
+    </div>
+  );
+};
+
+export default Not_Found;
